✨ :: 자리 정보 가져오도록 함수들 변경

+++ app/src/main/java/com/g3c1/oasis_android/remote/api/PurchaseApi.kt
@@ -14,5 +14,5 @@ interface PurchaseApi {
     suspend fun sendsTheOrderedFoodListToTheServer(@Body body: OrderedTableInfoDTO): Response<Unit>
 
     @GET("v2/purchase/{seatId}")
-    suspend fun getMyOrderInfo(@Path("seatId") seatId: Int): Response<RemoteOrderInfoDTO>
+    suspend fun getMyOrderInfo(@Path("seatId") seatId: Long): Response<RemoteOrderInfoDTO>
 }
